+++ permissions/LicensePlates/permissions.ps1
@@ -1,0 +1,8 @@
+$outputContext.Permissions.Add(
+    @{
+        DisplayName    = "LicensePlate"
+        Identification = @{
+            Reference = "LicensePlate"
+        }
+    }
+)