+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [1.0.2] - 29-07-2026
+
+Added functionality to use licenseplates as identifications
+
+### Added
+
+- Licenseplates as permission entitlement.
+- Create licenseplates as resources in Simac Pronto
+- Importscript for subpermissions
+
+### Changed
+
+### Deprecated
+
+### Removed
+
 ## [1.0.1] - 23-04-2026
 
 Some minor adjustments after the first release.
@@ -17,7 +33,6 @@ Some minor adjustments after the first release.
 ### Deprecated
 
 ### Removed
-
 
 ## [1.0.0] - 29-01-2026